Refactor ID fields to Long and update group handling
Changed various model ID fields from Int to Long for consistency and scalability. Updated Preferences to support Long values and refactored ViewModels and UI logic to use Long for group and chat IDs. Also improved group creation and login logic to handle GROUPLIST as a string and updated related repository calls.

+++ app/src/main/java/com/example/studify/data/model/AnnounceModel.kt
@@ -9,7 +9,7 @@ data class AnnounceModel(
     @SerializedName("result")      @Expose var result: List<AnnounceContent> = emptyList()
 ) {
     data class AnnounceContent(
-        @SerializedName("ANNOUNCE_ID")      @Expose var announceId: Int? = null,
+        @SerializedName("ANNOUNCE_ID")      @Expose var announceId: Long? = null,
         @SerializedName("ANNOUNCE_NAME")    @Expose var announceName: String? = "",
         @SerializedName("ANNOUNCE_CONTENT") @Expose var announceContent: String? = "",
         @SerializedName("ANNOUNCE_DATE")    @Expose var announceDate: String? = null,
