+++ app/src/main/java/com/example/studify/data/model/ChatModel.kt
@@ -12,8 +12,8 @@ data class ChatModel (
     @SerializedName("result") @Expose var list: List<StfChat>? =null
 ) : Serializable {
     data class  StfChat(
-        @SerializedName("ID") @Expose var ID: Int? = null,
-        @SerializedName("GROUPID") @Expose var GROUPID: Int? = null,
+        @SerializedName("ID") @Expose var ID: Long? = null,
+        @SerializedName("GROUPID") @Expose var GROUPID: Long? = null,
         @SerializedName("CHATNAME") @Expose var CHATNAME: String? = null,
         @SerializedName("CHAT") @Expose var CHAT: String? = null,
         @SerializedName("TIME") @Expose var TIME : Timestamp? = null
